+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Add `finalize` method for the Dispute API
 * Add `find` method for the Dispute API
 * Add `removeEvidence` method for the Dispute API
+* Add `search` method for the Dispute API
 
 ## 3.23.1
 * Fix token generator return type - thanks @jjok!